// client/src/app/models/Profile.ts
export interface Profile {
    codUsuario?: number;
    nomUsuario?: string;
    bio?: string;
    sexo?: string;
    alias?: string;
    email?: string;
    // fechaCreacion: Date;
    apellido1?: string;
    apellido2?: string;
    contrasena?: string;
    visibilidad?: string;
    icono?: string;
}